+++ README.md
@@ -209,12 +209,13 @@ Environment overrides: `MODEL_PATH`, `PTYPE`, `DTYPE`, `BYTE_ORDER`, `MAX_TOKENS
 
 | Module | Contents |
 |--------|----------|
+| `lora` | `LoraAdapter`, `LoraAdapterSet` (checkpoint I/O), `LoraAdamOptimizer` — pure Java, no GGUF/CUDA |
 | `api` | OpenAPI 3.0 spec, JAX-RS interfaces, `inference.proto` |
 | `registry` | `NodeDescriptor`, `ShardPlanner`, `ShardMap`, `ParallelismType`, `TensorShardAssignment`, `TensorShardPlanner` |
 | `coordinator` | `GenerationLoop`, `RequestScheduler`, `FaultTolerantPipeline`, Javalin REST, SSE |
 | `kvcache` | `KVCacheManager`, `GpuKVCache`, `CpuKVCache`, `PrefixCache` |
 | `tokenizer` | `GgufTokenizer` (SentencePiece BPE + GPT-2 BPE; auto-detected), `ChatTemplate`, `SimpleTokenizer`, `TokenizerEvent`, `TemplateFormatEvent` |
-| `node` | `LlamaTransformerHandler`, `Phi3TransformerHandler`, `ForwardPassHandlerLoader`, `EmbeddedNodeServer`, `NodeMain`, `NodeKVCacheAdapter`, `MatVec`, `CpuMatVec`, `CudaMatVec`, `GpuContext`, `DeviceFloatMatrix`, `DeviceHalfMatrix`, `GgufReader`, `LlamaConfig`, `ActivationCodec`, `ActivationBECodec`, `ActivationLECodec`, `ShardContext`, `TensorShardContext`, `LoraAdapter`, `LoraAdapterSet`, `LoraAdamOptimizer`, `LoraTrainableHandler`, `MatVecEvent`, `ForwardPassEvent`, `LoraTrainEvent` |
+| `node` | `LlamaTransformerHandler`, `Phi3TransformerHandler`, `ForwardPassHandlerLoader`, `EmbeddedNodeServer`, `NodeMain`, `NodeKVCacheAdapter`, `MatVec`, `CpuMatVec`, `CudaMatVec`, `GpuContext`, `DeviceFloatMatrix`, `DeviceHalfMatrix`, `GgufReader`, `LlamaConfig`, `ActivationCodec`, `ActivationBECodec`, `ActivationLECodec`, `ShardContext`, `TensorShardContext`, `LoraQvInitializer`, `LoraMerge`, `LoraTrainableHandler`, `MatVecEvent`, `ForwardPassEvent`, `LoraTrainEvent` — depends on `lora` for adapter tensors and Adam |
 | `sampler` | Temperature, top-k, top-p, repetition penalty — pure Java |
 | `health` | `NodeHealth` record (`nodeId`, `nodeRole`, `vramPressure`, `cpuLoad`, `inferenceLatencyP99Ms`, `throughputBytesPerSec`), `HealthReporter` (pushes 5 s snapshots via HTTP; `recordLatency()` for coordinator, `recordBytes()` for nodes), `HealthMain` (Javalin sidecar: `/health-ui` dashboard, `/health/probe`, `/health/circuits`), `HealthEvaluator`, `CircuitBreaker` |
 | `player` | `ConsoleMain` REPL (`cluster` / `local` / `lora` / `merge` commands), `ClusterHarness`, `ProcessPipelineClient`, `TensorParallelPipelineClient`, `LoraMergeMain` |
@@ -249,7 +250,7 @@ Template resolved from model type string via exact match then substring fallback
 ```bash
 mvn clean package -DskipTests          # build — produces shade jars
 
-mvn test -pl tokenizer,node,coordinator,sampler,kvcache,health,registry,player
+mvn test -pl tokenizer,lora,node,coordinator,sampler,kvcache,health,registry,player
                                        # unit tests — no model file, no GPU needed
 
 mvn verify -pl juno-master             # integration tests — forks 3 JVM nodes (stub mode)
@@ -338,7 +339,7 @@ Useful analysis patterns:
 - **Configurable activation byte order.** `ActivationCodec` is a zero-overhead static dispatcher: it reads `juno.byteOrder` once at class-load time and branches to either `ActivationBECodec` (big-endian, default) or `ActivationLECodec` (little-endian, native x86 order) for all encode/decode calls. The byte order is propagated consistently across every JVM in a cluster — `ClusterHarness` injects `-Djuno.byteOrder` into every forked node process; `juno-deploy.sh` writes it into `/etc/juno/node.env` for systemd-managed nodes. The cluster health endpoint reports the active byte order; the web console displays a live badge.
 - **KV cache wired to the node-level manager.** `NodeKVCacheAdapter` connects `LlamaTransformerHandler` and `Phi3TransformerHandler` to the `KVCacheManager` (GPU byte-budget LRU + Caffeine W-TinyLFU CPU tier). Every forward pass flushes key/value data write-through into both tiers. If a local entry is evicted under heap pressure, the next forward pass at that position restores it from the manager transparently. `evict(requestId)` propagates to both the local HashMap and both cache tiers, closing the gap that previously made the entire `kvcache` module inert at the node level.
 - **Star AllReduce for tensor parallel.** No inter-node communication. The coordinator collects partial logit vectors from all N nodes and sums them in O(N × vocabSize). Simpler than ring-AllReduce and requires no InfiniBand.
-- **LoRA fine-tuning without touching the base model.** `LoraTrainableHandler` wraps `LlamaTransformerHandler` and adds trainable low-rank adapters (A/B matrices, rank 4–16) on the Q and V projections. The frozen weights stay quantized at all times — backward passes dequantize one block per row via dedicated `transposedQ4K` / `transposedQ5K` / `transposedQ6K` scatter-reduce implementations. Adapters are persisted to a `.lora` binary checkpoint; the GGUF is never modified. To produce a standalone merged model, use `./juno merge` — see **Merge** below.
+- **LoRA fine-tuning without touching the base model.** Adapter tensors, checkpoint I/O, and Adam live in the **`lora`** Maven module (`cab.ml.juno.lora`). `LoraTrainableHandler` in **`node`** loads GGUF weights, applies adapters on Q/V projections, and runs training; forward matmuls use GPU (FP16 device weights + cuBLAS) when `CudaMatVec` is selected. Backward through frozen weights still dequantises one block per row on CPU. Adapters persist to a `.lora` file; the GGUF is never modified. For a standalone merged model, use `./juno merge`.
 - **Native LoRA merge (`juno merge`).** `LoraMerge` writes a new GGUF where the 44 LoRA-patched projection tensors (wq/wv on every layer) are stored as **F32** instead of being re-quantised. This is essential: the typical LoRA delta (~6×10⁻⁴ per element) is smaller than Q4_K quantisation noise (~3×10⁻³), so re-quantising would erase all training. The F32 patched tensors are read by `GgufReader` and `LlamaTransformerHandler` identically to any other F32 tensor — no special-casing in inference. All other tensors (norms, FFN, embeddings, output projection) are copied verbatim in their original quantised form. The merged file is larger than the source (~1 GB for TinyLlama 1.1B Q4_K_M vs 667 MB) but requires no `.lora` sidecar at inference time.
 - **GPT-2 BPE and SentencePiece BPE both supported.** `GgufTokenizer` reads `tokenizer.ggml.model` from GGUF metadata. Value `"gpt2"` activates the GPT-2 / tiktoken path (Llama 3+): space-prefixed words use Ġ (U+0120), and special control tokens are pre-split longest-first before BPE. Any other value (null / `"llama"` / `"llama2"`) uses the SentencePiece path (Llama 1/2, TinyLlama, Mistral, Gemma, Phi-3). No configuration required — detection is automatic at load time.
 - **AWS infrastructure scripted.** `scripts/aws/launcher.sh` is a credential wrapper. `juno-deploy.sh` is a unified cluster lifecycle script (replaces the earlier `juno-infra.sh` / `juno-infra-ft.sh`): hardware is auto-detected during bootstrap — GPU nodes install CUDA and set `JUNO_USE_GPU=true`, CPU nodes skip it. Commands: `setup | start | stop | teardown | status | scan-regions`. Coordinator can be co-located on node 1 (default, free) or launched as a separate t3.medium. State persisted to `~/.juno-deploy-state`; Ctrl+C auto-stops instances before exit. After `setup` the script bootstraps all nodes (~5 min), writes `cluster-nodes.env`, starts the coordinator, and enters the live cluster monitor showing per-node CPU / mem, health, and estimated cost. Web console served at `http://<coordinator>:8080` once cluster is healthy.
